### 2026-05-05 #302 @danielfromearth
earthaccess: Python Library for NASA Earthdata APIs
Submitting Author: (@danielfromearth)
All current maintainers: (@andypbarrett, @battistowx, @jrbourbeau, @itcarroll, @chuckwondo, @mfisher87, @danielfromearth, @jhkennedy, @betolink, @jules32, @JessicaS11, @asteiker, @Sherwin-14)
Package Name: "earthaccess"
One-Line Description of Package: "`earthaccess` is a python library to search for, and download or stream NASA Earth science data with just a few lines of code."
Repository Link: "https://github.com/earthaccess-dev/earthaccess"
Version submitted: "0.17.0"
Version accepted: TBD

---

- [x] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in future interactions in spaces supported by pyOpenSci should it be accepted.

`earthaccess` is an open-source Python library that unifies the discovery, authentication, and access steps for working with NASA Earth science data into a few lines of Python.  These steps have historically required researchers to navigate separate systems, data provider-specific APIs, and substantial boilerplate code. Built around NASA's Earth Observing System Data and Information System (EOSDIS), which distributes over 100 petabytes of satellite imagery, climate records, atmospheric measurements, and other geospatial datasets, `earthaccess` handles authenticating with NASA's Earthdata Login service, searching NASA's Common Metadata Repository, and retrieving data via HTTPS download, direct cloud access, or streaming into analysis-ready formats — so that scientists can move from question to analysis without wrestling with infrastructure.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

`earthaccess` falls under the "Data Retrieval" category because it focuses specifically on retrieving scientific datasets from NASA's Earth Observing System Data and Information System (EOSDIS), streamlining the NASA-specific authentication flows and cloud-vs-on-premises access logic that would otherwise require researchers to write substantial boilerplate code to go from logging in, to searching for data, to accessing the actual data.

The target audience includes Earth scientists, remote sensing researchers, climate modelers, hydrologists, ecologists, and any researcher or educator who needs to work with NASA Earth science data. Scientific applications span any workflow that begins with accessing NASA Earthdata — from downloading TEMPO air quality observations for atmospheric chemistry studies, to streaming SWOT satellite altimetry for hydrology research, to constructing applications that process thousands of data files for large-scale climate analyses.

Several Python packages address parts of this workflow: `python-cmr` wraps the CMR search API but does not handle authentication or data download; `harmony-py` provides server-side subsetting and reformatting but not direct data access; and `icepyx` provides specialized tools for ICESat-2 data only. `earthaccess` is unique in integrating search, authentication, and access (both HTTPS download and direct S3 streaming) into a single, mission-agnostic abstraction that works across NASA EOSDIS data holdings.

- [`currently being drafted`] The package contains a `paper.md` matching [JOSS's requirements][JossPaperRequirements] with a high-level description in the package root or in `inst/` by the time you wish to submit to JOSS.
- [x] The package is deposited in a long-term repository with the DOI: https://doi.org/10.5281/zenodo.8365009

*Note: JOSS accepts our review as theirs. You will NOT need to go through another full review. JOSS will only review your paper.md file. Be sure to link to this pyOpenSci issue when a JOSS issue is opened for your package. Also be sure to tell the JOSS editor that this is a pyOpenSci reviewed package once you reach this step. Please note that the PyOpenSci reviewers will not be reviewing the paper.md file*

**P.S.** Have feedback/comments about our review process? Leave a comment [on our discourse forum][Comments]
