### 2026-05-06 8392a9d
Fixed errors and violations from C Rule
