### 2026-05-07 ef0e86d
Fix FileExistsError during direct in-region downloads
Fixes #1331
