### 2026-05-08 1eb2bc0
Added C9 back to ignore list.
Co-authored-by: Matt Fisher <3608264+mfisher87@users.noreply.github.com>
