+++ pyproject.toml
@@ -214,6 +214,7 @@ ignore = [
   "B023",
   "B905",
   "BLE001",
+  "C901",
   "COM812",
   "D1",
   "D205",
