### 2026-05-08 afd8df6
Removed noqa as C9 is added back to ignore list.
Co-authored-by: Matt Fisher <3608264+mfisher87@users.noreply.github.com>
