### 2026-05-08 f700976
Enable A (flake8-builtins) ruleset and fix violations (#1330)
