### 2026-05-08 75b2c50
Fixed errors related PERF and COM ruleset
