### 2026-05-08 9c979c0
Enable C4 (flake8-comprehensions) ruleset and fix violations (#1332)
Co-authored-by: Matt Fisher <3608264+mfisher87@users.noreply.github.com>
