+++ noxfile.py
@@ -65,10 +65,10 @@ def integration_tests(session: nox.Session) -> None:
         "tests/integration",
         "-rxXs",  # Show provided reason in summary for (x)fail, (X)pass, and (s)kipped tests
         *session.posargs,
-        env=dict(
-            EARTHDATA_USERNAME=os.environ["EARTHDATA_USERNAME"],
-            EARTHDATA_PASSWORD=os.environ["EARTHDATA_PASSWORD"],
-        ),
+        env={
+            "EARTHDATA_USERNAME": os.environ["EARTHDATA_USERNAME"],
+            "EARTHDATA_PASSWORD": os.environ["EARTHDATA_PASSWORD"],
+        },
     )
 
 
