+++ pyproject.toml
@@ -211,10 +211,6 @@ ignore = [
   "B023",
   "B905",
   "BLE001",
-  "C402",
-  "C408",
-  "C416",
-  "C420",
   "C901",
   "COM812",
   "D1",
