### 2026-05-11 f9d380d
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
