+++ earthaccess/results.py
@@ -62,7 +62,11 @@ class CustomDict(dict):
 
     def _filter_related_links(self, link_type: str) -> list[str]:
         """Filter RelatedUrls from the UMM fields on CMR."""
-        return [link["URL"] for link in self["umm"].get("RelatedUrls",[]) if link.get("Type") == link_type]
+        return [
+            link["URL"]
+            for link in self["umm"].get("RelatedUrls", [])
+            if link.get("Type") == link_type
+        ]
 
 
 class DataCollection(CustomDict):
