### 2026-05-11 17c6dbd
Enable PERF (performance) and COM (flake8-commas) rulesets and fix violations (#1336)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
