### 2026-05-12 38b9208
Fixed errors and violations related to RUF ruleset
