+++ pyproject.toml
@@ -238,11 +238,6 @@ ignore = [
   "PT012",
   "PT017",
   "PT027",
-  "RUF002",
-  "RUF012",
-  "RUF015",
-  "RUF021",
-  "RUF022",
   "S101",
   "S108",
   "S113",
