### 2026-05-12 34e3ca2
Continue progress on creating results classes
* Add init method for base class
* Replace `len` calls with `query.hits` in anticipation of generator approach
* Update preview/repr methods also in an anticipation of generator approach
