### 2026-05-12 95cf656
add query attribute
