### 2026-05-12 56af9eb
remove len from query.hits call
