### 2026-05-12 fcbde26
WIP: sketch out what a services plugin API might look like
