### 2026-05-12 020f7f0
fix calls to hits and params in __repr__, __str__ and __repr_html__
