### 2026-05-12 06b3544
add return type for query
