### 2026-05-14 a18addd
Refactored strings inside exception by storing them in variables in auth.py and __init__.py and enabled EM ruleset
