### 2026-05-15 9756ecd
Refactored strings inside exceptions by storing them in variables in results.py, search.py, store.py and files in tests and virtual folders
