+++ earthaccess/store.py
@@ -371,14 +371,14 @@ class Store:
             a s3fs file instance
         """
         if self.auth is None:
-            raise ValueError(
-                "A valid Earthdata login instance is required to retrieve S3 credentials",
-            )
+            msg = "A valid Earthdata login instance is required to retrieve S3 credentials"
+            raise ValueError(msg)
         if not any([concept_id, daac, provider, endpoint]):
-            raise ValueError(
+            msg = (
                 "At least one of the concept_id, daac, provider or endpoint"
-                "parameters must be specified. ",
+                "parameters must be specified. "
             )
+            raise ValueError(msg)
 
         if concept_id is not None:
             provider = self._derive_concept_provider(concept_id)
@@ -449,7 +449,8 @@ class Store:
         """
         if hasattr(self, "_http_session"):
             return self._http_session
-        raise AttributeError("The requests session hasn't been set up yet.")
+        msg = "The requests session hasn't been set up yet."
+        raise AttributeError(msg)
 
     def open(
         self,
@@ -508,7 +509,8 @@ class Store:
         pqdm_kwargs: Mapping[str, Any] | None = None,
         open_kwargs: dict[str, Any] | None = None,
     ) -> list[Any]:
-        raise NotImplementedError("granules should be a list of DataGranule or URLs")
+        msg = "granules should be a list of DataGranule or URLs"
+        raise NotImplementedError(msg)
 
     @_open.register
     def _open_granules(
@@ -528,10 +530,8 @@ class Store:
         )
 
         if self.auth is None:
-            raise ValueError(
-                "A valid Earthdata login instance is required to retrieve credentials",
-            )
-
+            msg = "A valid Earthdata login instance is required to retrieve credentials"
+            raise ValueError(msg)
         if self.in_region:
             if granules[0].cloud_hosted:
                 access = "direct"
@@ -559,11 +559,12 @@ class Store:
                         open_kwargs=open_kwargs,
                     )
                 except Exception as e:
-                    raise RuntimeError(
+                    msg = (
                         "An exception occurred while trying to access remote files on S3. "
                         "This may be caused by trying to access the data outside the us-west-2 region."
-                        f"Exception: {traceback.format_exc()}",
-                    ) from e
+                        f"Exception: {traceback.format_exc()}"
+                    )
+                    raise RuntimeError(msg) from e
             else:
                 return self._open_urls_https(
                     url_mapping,
@@ -595,13 +596,13 @@ class Store:
             # TODO: method to derive the DAAC from url?
             provider = provider
         else:
-            raise ValueError(
-                f"Schema for {granules[0]} is not recognized, must be an HTTP or S3 URL",
+            msg: str = (
+                f"Schema for {granules[0]} is not recognized, must be an HTTP or S3 URL"
             )
+            raise ValueError(msg)
         if self.auth is None:
-            raise ValueError(
-                "A valid Earthdata login instance is required to retrieve S3 credentials",
-            )
+            msg = "A valid Earthdata login instance is required to retrieve S3 credentials"
+            raise ValueError(msg)
 
         url_mapping: Mapping[str, None] = dict.fromkeys(granules)
         if self.in_region and granules[0].startswith("s3"):
@@ -618,11 +619,12 @@ class Store:
                         open_kwargs=open_kwargs,
                     )
                 except Exception as e:
-                    raise RuntimeError(
+                    msg = (
                         "An exception occurred while trying to access remote files on S3. "
                         "This may be caused by trying to access the data outside the us-west-2 region. "
-                        f"Exception: {traceback.format_exc()}",
-                    ) from e
[…]
