+++ earthaccess/api.py
@@ -632,7 +632,7 @@ def get_s3_filesystem(
     else:
         msg = (
             "Invalid set of input arguments given. Please provide either "
-            "a valid result, an endpoint, a daac, or a provider.",
+            "a valid result, an endpoint, a daac, or a provider."
         )
         raise ValueError(msg)
     return session
@@ -650,8 +650,6 @@ def get_edl_token() -> str:
 def auth_environ() -> dict[str, str]:
     auth = earthaccess.__auth__
     if not auth.authenticated:
-        msg = (
-            "`auth_environ()` requires you to first authenticate with `earthaccess.login()`",
-        )
+        msg = "`auth_environ()` requires you to first authenticate with `earthaccess.login()`"
         raise RuntimeError(msg)
     return {"EARTHDATA_USERNAME": auth.username, "EARTHDATA_PASSWORD": auth.password}
