### 2026-05-15 41551a0
Remove unnecessary type annotation
