### 2026-05-15 c3e7934
Enable EM (flake8-errmsg) ruleset and fix violations. (#1342)
Co-authored-by: Matt Fisher <3608264+mfisher87@users.noreply.github.com>
