+++ earthaccess/search.py
@@ -163,7 +163,8 @@ class DataCollections(CollectionQuery):
             TypeError: `doi` is not of type `str`.
         """
         if not isinstance(doi, str):
-            raise TypeError("doi must be of type str")
+            msg = "doi must be of type str"
+            raise TypeError(msg)
 
         self.params["doi"] = doi
         return self
@@ -185,7 +186,8 @@ class DataCollections(CollectionQuery):
             TypeError: `instrument` is not of type `str`.
         """
         if not isinstance(instrument, str):
-            raise TypeError("instrument must be of type str")
+            msg = "instrument must be of type str"
+            raise TypeError(msg)
 
         self.params["instrument"] = instrument
         return self
@@ -208,7 +210,8 @@ class DataCollections(CollectionQuery):
             TypeError: `project` is not of type `str`.
         """
         if not isinstance(project, str):
-            raise TypeError("project must be of type str")
+            msg = "project must be of type str"
+            raise TypeError(msg)
 
         self.params["project"] = project
         return self
@@ -240,7 +243,8 @@ class DataCollections(CollectionQuery):
         for key, val in kwargs.items():
             # verify the key matches one of our methods
             if key not in methods:
-                raise ValueError(f"Unknown key {key}")
+                msg = f"Unknown key {key}"
+                raise ValueError(msg)
 
             # call the method
             if isinstance(val, tuple):
@@ -297,7 +301,8 @@ class DataCollections(CollectionQuery):
             self
         """
         if has_granules is not None and not isinstance(has_granules, bool):
-            raise TypeError("has_granules must be of type bool or None")
+            msg = "has_granules must be of type bool or None"
+            raise TypeError(msg)
 
         if has_granules is not None:
             self.params["has_granules"] = has_granules
@@ -324,7 +329,8 @@ class DataCollections(CollectionQuery):
             TypeError: `cloud_hosted` is not of type `bool`.
         """
         if not isinstance(cloud_hosted, bool):
-            raise TypeError("cloud_hosted must be of type bool")
+            msg = "cloud_hosted must be of type bool"
+            raise TypeError(msg)
 
         self.params["cloud_hosted"] = cloud_hosted
         if hasattr(self, "DAAC"):
@@ -513,7 +519,8 @@ class DataGranules(GranuleQuery):
         for key, val in kwargs.items():
             # verify the key matches one of our methods
             if key not in methods:
-                raise ValueError(f"Unknown key {key}")
+                msg = f"Unknown key {key}"
+                raise ValueError(msg)
 
             # call the method
             if isinstance(val, tuple):
@@ -607,7 +614,8 @@ class DataGranules(GranuleQuery):
             TypeError: `cloud_hosted` is not of type `bool`.
         """
         if not isinstance(cloud_hosted, bool):
-            raise TypeError("cloud_hosted must be of type bool")
+            msg = "cloud_hosted must be of type bool"
+            raise TypeError(msg)
 
         if "short_name" in self.params:
             provider = find_provider_by_shortname(
@@ -636,9 +644,8 @@ class DataGranules(GranuleQuery):
             TypeError: if `granule_name` is not of type `str` or `Iterable[str]`.
         """
         if not isinstance(granule_name, Iterable):
-            raise TypeError(
-                "granule_name must be of type string or Iterable of strings",
-            )
+            msg = "granule_name must be of type string or Iterable of strings"
+            raise TypeError(msg)
         if not isinstance(granule_name, str):
             # Convert iterable to list of strings. Since str is also Iterable, make
             # sure we don't do this when granule_name is a string, otherwise
