### 2026-05-16 0775668
Enabled PGH ruleset and fixed violations
