+++ earthaccess/store.py
@@ -48,7 +48,7 @@ def _is_interactive() -> bool:
     Interactive sessions include Jupyter Notebooks, IPython REPL, and default Python REPL.
     """
     try:
-        from IPython import get_ipython  # type: ignore
+        from IPython import get_ipython  # type: ignore[import-not-found]
 
         # IPython Notebook or REPL:
         if get_ipython() is not None:
@@ -152,7 +152,7 @@ def _open_files(
         open_kw.setdefault("block_size", default_block_size)
 
         f = fs.open(url, **open_kw)
-        return EarthAccessFile(f, granule)  # type: ignore
+        return EarthAccessFile(f, granule)  # type: ignore[arg-type]
 
     # this {#n_jobs} is for the unittests as this method is not public and pqdm will have values at this point
     return pqdm(
@@ -889,7 +889,7 @@ class Store:
             None
         """
         if not hasattr(self.thread_locals, "local_thread_session"):
-            local_thread_session = self.auth.get_session()  # type: ignore
+            local_thread_session = self.auth.get_session()
             local_thread_session.headers.update(original_session.headers)
             local_thread_session.cookies.update(original_session.cookies)
             local_thread_session.auth = original_session.auth
