+++ tests/unit/test_store.py
@@ -92,7 +92,7 @@ class TestEula(unittest.TestCase):
             store.get([mocked_url], "/tmp")
 
     def tearDown(self):
-        self.auth = None
+        self.auth = None  # type: ignore[assignment]
 
     @responses.activate
     def test_store_can_create_https_fsspec_session(self):
@@ -132,7 +132,7 @@ class TestStoreSessions(unittest.TestCase):
         self.assertEqual(self.auth.token, json_response)
 
     def tearDown(self):
-        self.auth = None
+        self.auth = None  # type: ignore[assignment]
 
     @responses.activate
     def test_store_can_create_https_fsspec_session(self):
@@ -179,7 +179,7 @@ class TestStoreSessions(unittest.TestCase):
             if "s3-credentials" in daac:
                 responses.add(
                     responses.GET,
-                    daac["s3-credentials"],
+                    daac["s3-credentials"],  # type: ignore[arg-type]
                     json=mock_creds,
                     status=200,
                 )
@@ -192,7 +192,7 @@ class TestStoreSessions(unittest.TestCase):
 
         store = Store(self.auth)
         self.assertTrue(isinstance(store.auth, Auth))
-        for daac in [
+        for daac in [  # type: ignore[assignment]
             "NSIDC",
             "PODAAC",
             "LPDAAC",
@@ -202,7 +202,7 @@ class TestStoreSessions(unittest.TestCase):
             "OBDAAC",
             "ASDC",
         ]:
-            s3_fs = store.get_s3_filesystem(daac=daac)
+            s3_fs = store.get_s3_filesystem(daac=daac)  # type: ignore[arg-type]
             assert isinstance(s3_fs, s3fs.S3FileSystem)
             assert s3_fs.storage_options == expected_storage_options
 
@@ -302,7 +302,7 @@ class TestStoreSessions(unittest.TestCase):
                         side_effect=mock_download_file,
                     ):
                         # Test multi-threaded download
-                        pqdm(urls, store._download_file, n_jobs=n_threads)  # type: ignore
+                        pqdm(urls, store._download_file, n_jobs=n_threads)  # type: ignore[arg-type]
 
                 # We make sure we reuse the token up to N threads
                 self.assertTrue(len(cloned_sessions) <= n_threads)
@@ -314,7 +314,7 @@ class TestStoreSessions(unittest.TestCase):
 def test_earthaccess_file_getattr():
     fs = fsspec.filesystem("memory")
     with fs.open("foo", "wb") as f:
-        earthaccess_file = EarthAccessFile(f, granule="foo")
+        earthaccess_file = EarthAccessFile(f, granule="foo")  # type: ignore[arg-type]
         assert f.tell == earthaccess_file.tell
     fs.store.clear()
 
