+++ pyproject.toml
@@ -223,7 +223,6 @@ ignore = [
   "FIX002",
   "FURB122",
   "INP001",
-  "PGH003",
   "PLC0415",
   "PLR0913",
   "PLW0127",
