### 2026-05-16 6ee2ed3
Enabled PLW ruleset and fixed violations
