+++ pyproject.toml
@@ -226,8 +226,6 @@ ignore = [
   "PGH003",
   "PLC0415",
   "PLR0913",
-  "PLW0127",
-  "PLW0602",
   "PT006",
   "PT007",
   "PT009",
