### 2026-05-16 ddc0f40
Enabled FUR ruleset
