### 2026-05-16 6912353
Enable PLC ruleset and fixed violations
