+++ earthaccess/virtual/core.py
@@ -222,7 +222,7 @@ def _open_virtual_mfdataset(
 ) -> xr.Dataset:
     """Thin wrapper around ``vz.open_virtual_mfdataset`` for testability."""
     try:
-        import virtualizarr as vz
+        import virtualizarr as vz  # noqa: PLC0415
     except ImportError as exc:
         raise ImportError(
             "earthaccess.virtualize() requires `pip install earthaccess[virtualizarr]`",
@@ -265,7 +265,7 @@ def _load_via_kerchunk(
     Needed until https://github.com/zarr-developers/VirtualiZarr/issues/360
     is resolved. TODO: make sure this holds, I think this may have been resolved already.
     """
-    import xarray as xr
+    import xarray as xr  # noqa: PLC0415
 
     fs = earthaccess.get_fsspec_https_session()
 
