+++ tests/unit/test_virtual.py
@@ -16,6 +16,15 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 from earthaccess.results import DataCollection, DataGranule
+from earthaccess.virtual._credentials import (
+    get_granule_credentials_endpoint_and_region,
+)
+from earthaccess.virtual._parser import (
+    SUPPORTED_PARSERS,
+    get_urls_for_parser,
+    resolve_parser,
+)
+from earthaccess.virtual.core import virtualize
 
 # ---------------------------------------------------------------------------
 # Helpers
@@ -62,16 +71,12 @@ def _patch_internals(mock_vds: MagicMock | None = None):
 
 def test_virtualize_empty_granules_raises() -> None:
     """virtualize() raises ValueError when granules list is empty."""
-    from earthaccess.virtual.core import virtualize
-
     with pytest.raises(ValueError, match=r"[Nn]o granules"):
         virtualize([])
 
 
 def test_virtualize_multi_granule_no_concat_dim_raises() -> None:
     """virtualize() raises ValueError for >1 granule without concat_dim."""
-    from earthaccess.virtual.core import virtualize
-
     with (
         patch(
             "earthaccess.virtual.core.build_obstore_registry",
@@ -84,8 +89,6 @@ def test_virtualize_multi_granule_no_concat_dim_raises() -> None:
 
 def test_virtualize_invalid_parser_string_raises() -> None:
     """virtualize() raises ValueError for an unrecognised parser string."""
-    from earthaccess.virtual.core import virtualize
-
     with pytest.raises(ValueError, match="BadParser"):
         virtualize(_make_granules(1), parser="BadParser")
 
@@ -97,8 +100,6 @@ def test_virtualize_invalid_parser_string_raises() -> None:
 
 def test_virtualize_load_false_returns_virtual_dataset() -> None:
     """virtualize(load=False) returns the raw virtual dataset without calling kerchunk."""
-    from earthaccess.virtual.core import virtualize
-
     mock_vds = MagicMock()
     reg_patch, open_patch = _patch_internals(mock_vds)
     with (
@@ -114,8 +115,6 @@ def test_virtualize_load_false_returns_virtual_dataset() -> None:
 
 def test_virtualize_load_true_delegates_to_kerchunk(tmp_path) -> None:
     """virtualize(load=True) calls _load_via_kerchunk and returns its result."""
-    from earthaccess.virtual.core import virtualize
-
     expected_ds = MagicMock()
     reg_patch, open_patch = _patch_internals()
     with (
@@ -143,8 +142,6 @@ def test_virtualize_load_true_delegates_to_kerchunk(tmp_path) -> None:
 
 def test_virtualize_dmrpp_fallback_emits_user_warning() -> None:
     """When DMR++ sidecars are missing, virtualize() warns and retries with HDFParser."""
-    from earthaccess.virtual.core import virtualize
-
     mock_vds_hdf = MagicMock()
     call_count = {"n": 0}
 
@@ -178,8 +175,6 @@ def test_virtualize_dmrpp_fallback_emits_user_warning() -> None:
 
 def test_supported_parsers_contains_canonical_names() -> None:
     """SUPPORTED_PARSERS includes the three canonical parser names."""
-    from earthaccess.virtual._parser import SUPPORTED_PARSERS
-
     assert isinstance(SUPPORTED_PARSERS, frozenset)
     assert {"DMRPPParser", "HDFParser", "NetCDF3Parser"} <= SUPPORTED_PARSERS
 
@@ -191,15 +186,11 @@ def test_supported_parsers_contains_canonical_names() -> None:
 
 def test_resolve_parser_dmrpp_returns_instance() -> None:
     """resolve_parser('DMRPPParser') returns a DMRPPParser instance."""
[…]
