+++ pyproject.toml
@@ -227,7 +227,6 @@ ignore = [
   "FURB122",
   "INP001",
   "PGH003",
-  "PLC0415",
   "PLR0913",
   "PLW0127",
   "PLW0602",
