### 2026-05-16 7f3e3b4
Enabled ARG ruleset
