+++ pyproject.toml
@@ -204,8 +204,6 @@ ignore = [
   "ANN202",
   "ANN204",
   "ANN401",
-  "ARG001",
-  "ARG002",
   "B018",
   "B019",
   "B023",
@@ -253,7 +251,7 @@ combine-as-imports = true
 
 [tool.ruff.lint.per-file-ignores]
 # Allow `print` and `pprint` in notebooks
-"*.ipynb" = ["T201", "T203", "N803", "RET504", "ERA001", "PLR2004", "PD002", "PD010", "PD011", "EM"]
+"*.ipynb" = ["T201", "T203", "N803", "RET504", "ERA001", "PLR2004", "PD002", "PD010", "PD011", "EM", "ARG001"]
 "tests/**/*.py" = ["PLR2004"]
 
 [tool.ruff.lint.pydocstyle]
