### 2026-05-16 b1ad435
Enable PLR ruleset and fixed violations
