+++ pyproject.toml
@@ -223,7 +223,6 @@ ignore = [
   "INP001",
   "PGH003",
   "PLC0415",
-  "PLR0913",
   "PLW0127",
   "PLW0602",
   "PT006",
