### 2026-05-17 420ae5a
Enable ARG (flake8-unused-arguments)  and PLR (Refactor) rulesets (#1347)
