### 2026-05-18 10c1228
Inverted if else logic for the _open_urls function in store.py
