### 2026-05-18 7615504
Revert "[pre-commit.ci] auto fixes from pre-commit.com hooks"
This reverts commit c457264705d4c19c8e37502b6fcc152beb9afccf.
