### 2026-05-18 d460cfb
Re-un-ignore PLR ruleset
