+++ pyproject.toml
@@ -222,7 +222,6 @@ ignore = [
   "INP001",
   "PGH003",
   "PLC0415",
-  "PLR0913",
   "PT006",
   "PT007",
   "PT009",
