### 2026-05-18 019b1da
fix typing defs
