### 2026-05-18 06f4fb7
fix mypy for manifest patching
