### 2026-05-18 bb16701
override direct access for virtual stores, todo: find a way to test in region access from gh actions
