### 2026-05-18 5148ff9
Enable PLW (Warning) and FURB (refurb) rulesets and fix violations (#1345)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Matt Fisher <mfisher87@gmail.com>
