+++ pyproject.toml
@@ -219,12 +219,9 @@ ignore = [
   "F401",
   "FBT",
   "FIX002",
-  "FURB122",
   "INP001",
   "PGH003",
   "PLC0415",
-  "PLW0127",
-  "PLW0602",
   "PT006",
   "PT007",
   "PT009",
