### 2026-05-18 25c4066
Enable PLC (Convention) ruleset and fix violations (#1346)
Co-authored-by: Matt Fisher <3608264+mfisher87@users.noreply.github.com>
