+++ pyproject.toml
@@ -221,7 +221,6 @@ ignore = [
   "FIX002",
   "INP001",
   "PGH003",
-  "PLC0415",
   "PT006",
   "PT007",
   "PT009",
