### 2026-05-18 4d3b57d
update warning rules
