+++ earthaccess/virtual/_credentials.py
@@ -111,11 +111,11 @@ def build_obstore_registry(
         ImportError: If ``earthaccess[virtualizarr]`` is not installed.
     """
     try:
-        from obstore.auth.earthdata import (  # noqa: PLC0415
+        from obstore.auth.earthdata import (
             NasaEarthdataCredentialProvider,
         )
-        from obstore.store import HTTPStore, S3Store  # noqa: PLC0415
-        from virtualizarr.registry import ObjectStoreRegistry  # noqa: PLC0415
+        from obstore.store import HTTPStore, S3Store
+        from virtualizarr.registry import ObjectStoreRegistry
     except ImportError:
         msg = (
             "earthaccess.virtualize() requires `pip install earthaccess[virtualizarr]`"
