+++ pyproject.toml
@@ -62,10 +62,7 @@ Changelog = "https://github.com/earthaccess-dev/earthaccess/blob/main/CHANGELOG.
 [project.optional-dependencies]
 # IMPORTANT: When adding a new optional dependencies group, include it in this
 # "all" optional dependencies group.
-all = [
-  "earthaccess[kerchunk]",
-  "earthaccess[virtualizarr]",
-]
+all = ["earthaccess[kerchunk]", "earthaccess[virtualizarr]"]
 
 kerchunk = [
   "kerchunk>=0.2.9",
@@ -78,7 +75,7 @@ kerchunk = [
 virtualizarr = [
   "numpy >=1.26.4",
   "zarr >=3.1.1",
-  "virtualizarr",
+  "virtualizarr >=2.1.2",
   "icechunk",
   "kerchunk>=0.2.9",
   "dask",
@@ -92,12 +89,7 @@ virtualizarr = [
 ]
 
 [dependency-groups]
-dev = [
-  "bump-my-version >=1.3.0",
-  "nox",
-  "pre-commit >=2.4",
-  "ruff >=0.6.3",
-]
+dev = ["bump-my-version >=1.3.0", "nox", "pre-commit >=2.4", "ruff >=0.6.3"]
 test = [
   "zarr >=3.1.1",
   "numpy >=1.26.4",
@@ -254,7 +246,19 @@ combine-as-imports = true
 
 [tool.ruff.lint.per-file-ignores]
 # Allow `print` and `pprint` in notebooks
-"*.ipynb" = ["T201", "T203", "N803", "RET504", "ERA001", "PLR2004", "PD002", "PD010", "PD011", "EM", "ARG001"]
+"*.ipynb" = [
+  "T201",
+  "T203",
+  "N803",
+  "RET504",
+  "ERA001",
+  "PLR2004",
+  "PD002",
+  "PD010",
+  "PD011",
+  "EM",
+  "ARG001",
+]
 "tests/**/*.py" = ["PLR2004"]
 
 [tool.ruff.lint.pydocstyle]
