### 2026-05-19 bfb5170
parallelize virtualizarr tests
