+++ pyproject.toml
@@ -97,6 +97,7 @@ test = [
   "pytest >=8.3",
   "pytest-cov >=5.0",
   "pytest-watch >=4.2",
+  "pytest-xdist >=3.8",
   "python-magic >=0.4; platform_system != 'Windows'",
   "python-magic-bin >=0.4; platform_system == 'Windows'",
   "responses >=0.14",
