+++ .github/workflows/integration-test.yml
@@ -60,13 +60,21 @@ jobs:
       matrix:
         include:
           - python-version: 3.12
+            test-group: virtualizarr
             # We must save the PR number only once, and this also prevents
             # "duplicate artifact" errors caused when attempting to save the same
             # artifact for every python version we're running with.  More specifically,
             # save-pr-number must be set to 'true' for 1 (and only 1) matrix combo
             # (it doesn't matter which combo).
             save-pr-number: true
+          - python-version: 3.12
+            test-group: other
+            save-pr-number: false
           - python-version: 3.13
+            test-group: virtualizarr
+            save-pr-number: false
+          - python-version: 3.13
+            test-group: other
             save-pr-number: false
       fail-fast: false
 
@@ -142,15 +150,28 @@ jobs:
           EARTHDATA_USERNAME: ${{ secrets.EDL_USERNAME }}
           EARTHDATA_PASSWORD: ${{ secrets.EDL_PASSWORD }}
         run: |
-          # -rxXs: Show provided (r)eason in summary for (x)fail, (X)pass, and (s)kipped tests
-          uv run pytest tests/integration \
-            -rxXs \
-            --cov=earthaccess \
-            --cov-report=term-missing \
-            --capture=no \
-            --color=yes \
-            --tb=native \
-            --log-cli-level=INFO
+          if [[ "${{ matrix.test-group }}" == "virtualizarr" ]]; then
+            uv run pytest tests/integration/test_virtualizarr.py \
+              -n logical \
+              -rxXs \
+              --cov=earthaccess \
+              --cov-report=term-missing \
+              --capture=no \
+              --color=yes \
+              --tb=native \
+              --log-cli-level=INFO
+          else
+            uv run pytest tests/integration/ \
+              --ignore=tests/integration/test_virtualizarr.py \
+              -n logical \
+              -rxXs \
+              --cov=earthaccess \
+              --cov-report=term-missing \
+              --capture=no \
+              --color=yes \
+              --tb=native \
+              --log-cli-level=INFO
+          fi
 
       - name: Upload coverage report
         # Don't upload coverage when using the `act` tool to run the workflow locally
