### 2026-05-19 56609bc
merging kerchunk into virtualizarr deps
