### 2026-05-20 2eeb773
Fix test deprecation warning for store (#1351)
Mock get_s3_filesystem to avoid real network calls in CI.
Suppress ruff SIM117 by combining nested with statements.
