### 2026-05-28 9aaa01f
docs: surface a Get help section on the docs home page
