### 2026-05-28 #302 @kysolvik
Hi @danielfromearth!

My name is Kylen, I'm a member of the EiC team here at pyOpenSci. Thanks for you submission of `earthaccess`! It looks like a great package and a good fit for pyOpenSci. Next step is[ pre-review checks](https://www.pyopensci.org/software-peer-review/how-to/editor-in-chief-guide.html#editor-in-chief-checklist), which I'll be tackling in the coming days. Thanks for your patience, I'll be back in touch asap!
