### 2026-06-05 9faeb78
Navigation structure changes
Building off the conversation here: https://github.com/earthaccess-dev/earthaccess/issues/1200 
I made some changes to the navigation to categorize content into more diataxis-esque groupings. Aside from that, we were inconsistent in our capitalization, so I tried to address that as well. Let me know if this is in line with what you two were thinking, @mfisher87 and @asteiker !
