### 2026-06-09 684d6a9
Edit navigation structure in .nav.yml
After discussion, moved "How-to" content above "Explanation" and removed the "s" from "How-tos".
