### 2026-06-09 3c417b7
Clarify community manager role in GH project
