### 2026-06-09 6fa0e8a
Remove issue vs discussion content
