+++ docs/contributor/triaging.md
@@ -8,16 +8,24 @@ This document outlines our approach to triaging issues on GitHub, including guid
 
 1. A [new issue](https://github.com/earthaccess-dev/earthaccess/issues/new/choose) is created, either using a pre-existing template, or as a blank issue.
 2. Issue triaging is led by the `earthaccess` community manager, though any community member is welcome and encouraged to contribute. Triaging involves:
-   * Determining whether the issue should be worked or not. If yes, move to "Backlog" status, otherwise close as not planned.
-   * Adding or adjusting issue labels.
-   * Adding an issue prioritization. 
-   * Responding and follow up as needed (i.e. tagging relevant earthaccess maintainers for further support).
-3. Issues are reviewed and groomed by the `earthaccess` community manager using the `earthaccess` [GitHub Project](https://github.com/orgs/earthaccess-dev/projects/1). 
-4. An issue is worked following the [Pull Request (PR) Guide](./pr-guide.md).
-   * [Assign](https://docs.github.com/en/issues/tracking-your-work-with-issues/using-issues/assigning-issues-and-pull-requests-to-other-github-users) the issue to the implementer.
-   * Update status to `In Progress`.
-   * Link the issue to the related PR with a comment "Resolves #N" (`N` is the issue number). The issue will auto-close when the PR is merged.
-   * If the previous step is missed, [link any related PRs](https://docs.github.com/en/issues/tracking-your-work-with-issues/using-issues/linking-a-pull-request-to-an-issue) before manually closing an issue.
+
+    * Determining whether the issue should be worked or not. If yes, move to "Backlog" status, otherwise close as not planned.
+
+    * Adding or adjusting issue labels.
+
+    * Adding an issue prioritization.
+
+    * Responding and follow up as needed (i.e. tagging relevant earthaccess maintainers for further support).
+4. Issues are reviewed and groomed by the `earthaccess` community manager using the `earthaccess` [GitHub Project](https://github.com/orgs/earthaccess-dev/projects/1). 
+5. An issue is worked following the [Pull Request (PR) Guide](./pr-guide.md).
+
+    * [Assign](https://docs.github.com/en/issues/tracking-your-work-with-issues/using-issues/assigning-issues-and-pull-requests-to-other-github-users) the issue to the implementer.
+
+    * Update status to `In Progress`.
+
+    * Link the issue to the related PR with a comment "Resolves #N" (`N` is the issue number). The issue will auto-close when the PR is merged.
+
+    * If the previous step is missed, [link any related PRs](https://docs.github.com/en/issues/tracking-your-work-with-issues/using-issues/linking-a-pull-request-to-an-issue) before manually closing an issue.
 
 Details on each of these workflow steps are provided below. 
 
