### 2026-06-09 f097fcc
Ruff format fix
Co-authored-by: Matt Fisher <mfisher87@users.noreply.github.com>
