### 2026-06-09 a967bee
Results subclasses must pass query to super init
