### 2026-06-09 6590d9a
Address ruff checks
