### 2026-06-09 52a142d
Refine items type annotation in Result initializer
