### 2026-06-09 6d96ee2
Revise user guide landing content
Added an info note about the User guide content structure, did some word-smithing of the content, and added some formatting.
