+++ docs/user/index.md
@@ -1,13 +1,24 @@
 # User guide
 
-`earthaccess` enables authentication, search and access for NASA Earth science data.  The package enables users to find, and download or stream data archived both at NASA DAACs and in the AWS-hosted Earthdata Cloud.  The User Guide introduces these three functions.
+!!! note "This documentation follows the Diátaxis framework:"
 
-The [Authentication](./authenticate.md) section shows how `earthaccess` can be used to manage Earthdata Login and also Earthdata Cloud credentials.
+    - **How-to guides** — goal-oriented instructions for specific tasks.
+    - **Explanation** — understanding-oriented discussions of concepts.
+    - **Tutorials** — guided learning-oriented lessons.
+    - **Reference** — information-oriented technical details.
+    
+    Learn more about Diátaxis at <https://diataxis.fr/>.
+
+**`earthaccess` enables authentication, search, and access for NASA Earth science data.**
+
+The package enables users to find, download, or stream data archived both at NASA DAACs and in the AWS-hosted Earthdata Cloud. The User guide contains How-to guides, Tutorials, Explanations, and Reference content to help users learn about and use `earthaccess`.
+
+Navigate to [Quick Start](./quick-start.md) to get working with `earthaccess` as quickly as possible. 
+
+**To learn more about the primary functions:**
+
+The [Authentication](./authenticate.md) section shows how `earthaccess` can be used to manage Earthdata Login and Earthdata Cloud credentials.
 
 The [Search](./search.md) section introduces how to search for NASA datasets (collections) and specific data granules.
 
 The [Access](./access.md) section introduces how to use `earthaccess` to download data to a local machine or stream data by opening it directly in memory.
-
-Additional usage information can be found in the User Reference section.  The How-To sections contains short code examples for specific workflows.  Longer tutorials for working with particular datasets are in the Tutorials section.
-
-The Glossary section contains definitions for NASA terminology and Cloud computing.
