### 2026-06-11 9a20611
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
