### 2026-06-11 5d8ec2f
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
