+++ docs/.nav.yml
@@ -66,4 +66,3 @@ nav:
           - "Integration tests": "contributor/integration-tests.md"
       - Reference:
           - "Naming conventions": "contributor/naming-convention.md"
-
