### 2026-06-12 ecd63bf
Fix badge job: force Node.js 24 for deprecated node20 action
