### 2026-06-12 8d6b0c8
Pin transitive aiohttp testing dependency
Works around vcrpy issue https://github.com/kevin1024/vcrpy/issues/995

**This is intended to be temporary. This commit should be reverted once
vcrpy releases, and vcrpy's lower bound should be bumped to that new
release.**
