### 2026-06-12 1db90f1
Lower-bound aiohttp so lowest-direct dependency resolution doesn't pull in v0.1
