### 2026-06-19 67feb1f
Updates per code review: an affiliation, HTTP access, and community
Co-authored-by: Joseph H Kennedy <me@jhkennedy.org>
