### 2026-06-19 4b25443
Add references for earthaccess-dependents
