### 2026-06-19 b2d362e
add bibtex entries for earthaccess-dependent references
