+++ paper/paper.bib
@@ -160,3 +160,36 @@
   url = {https://opendap.github.io/DMRpp-wiki/DMRpp.html},
   note = {Accessed: 2026-03-05}
 }
+
+@article{Scheick2025,
+  title = {Software to Enable Ocean Discoveries: A Case Study With {ICESat-2} and {Argo}},
+  author = {Scheick, J. and Piunno, R. and Fair, Z. and Tilling, R. and Di Bella, A. and Abib, N. and Bisson, K. M.},
+  journal = {Geoscience Data Journal},
+  volume = {12},
+  number = {3},
+  year = {2025},
+  publisher = {Wiley},
+  doi = {10.1002/gdj3.291}
+}
+
+@article{Widlansky2025,
+  title = {Building an Intelligent Data Exploring Assistant for Geoscientists},
+  author = {Widlansky, Matthew J. and Komar, Nemanja},
+  journal = {Journal of Geophysical Research: Machine Learning and Computation},
+  volume = {2},
+  number = {3},
+  year = {2025},
+  publisher = {American Geophysical Union (AGU)},
+  doi = {10.1029/2025JH000649}
+}
+
+@article{Xiahou2026,
+  title = {Hydrographic Variability in the Central {Ross Ice Shelf} Cavity and the Implications for Ocean Circulation},
+  author = {Xiahou, Yingpu and Brewer, Mike and Hulbe, Christina and Stevens, Craig},
+  journal = {Journal of Geophysical Research: Oceans},
+  volume = {131},
+  number = {1},
+  year = {2026},
+  publisher = {American Geophysical Union (AGU)},
+  doi = {10.1029/2025JC023511}
+}
