### 2026-06-20 d96415a
Replace aiohttp pin with vcrpy fixed version
