### 2026-06-23 fce8a0a
Update navigation structure and rename topics
