### 2026-06-23 ac9d54c
Update links in user guide for explanation directory
