+++ docs/user/index.md
@@ -2,11 +2,11 @@
 
 `earthaccess` enables authentication, search and access for NASA Earth science data.  The package enables users to find, and download or stream data archived both at NASA DAACs and in the AWS-hosted Earthdata Cloud.  The User Guide introduces these three functions.
 
-The [Authentication](./authenticate.md) section shows how `earthaccess` can be used to manage Earthdata Login and also Earthdata Cloud credentials.
+The [Authentication](./explanation/authenticate.md) section shows how `earthaccess` can be used to manage Earthdata Login and also Earthdata Cloud credentials.
 
-The [Search](./search.md) section introduces how to search for NASA datasets (collections) and specific data granules.
+The [Search](./explanation/search.md) section introduces how to search for NASA datasets (collections) and specific data granules.
 
-The [Access](./access.md) section introduces how to use `earthaccess` to download data to a local machine or stream data by opening it directly in memory.
+The [Access](./explanation/access.md) section introduces how to use `earthaccess` to download data to a local machine or stream data by opening it directly in memory.
 
 Additional usage information can be found in the User Reference section.  The How-To sections contains short code examples for specific workflows.  Longer tutorials for working with particular datasets are in the Tutorials section.
 
