+++ docs/user/explanation/backwards-compatibility.md
@@ -15,7 +15,7 @@ We use [Semantic Versioning (SemVer)](https://semver.org/) to tell users what to
 
 ### Versioning
 
-We will follow SemVer. All version changes will consider [this to be the public API documentation](../api/index.md) for the purposes of deciding whether a change is breaking or non-breaking.
+We will follow SemVer. All version changes will consider [this to be the public API documentation](../../api/index.md) for the purposes of deciding whether a change is breaking or non-breaking.
 
 ### Release Communication
 
