### 2026-06-24 72bc1cb
Edit user guide index content and update section links
