### 2026-06-25 d1f5a34
Update user guide introduction and structure
