### 2026-06-25 1e99c0a
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
