+++ docs/user/index.md
@@ -2,7 +2,7 @@
 
 **`earthaccess` enables authentication, search, and access for NASA Earth science data.**
 
-This user guide contains different forms of content for various user needs. Navigate to [Quick Start](./quick-start.md) to get working with `earthaccess` as quickly as possible. 
+This user guide contains different forms of content for various user needs. Navigate to [Quick Start](./quick-start.md) to get working with `earthaccess` as quickly as possible.
 
 !!! note "This documentation follows the Diátaxis framework:"
 
@@ -10,7 +10,7 @@ This user guide contains different forms of content for various user needs. Navi
     - **Tutorials** — longer, guided learning-oriented lessons.
     - **Explanation** — understanding-oriented discussions of concepts.
     - **Reference** — information-oriented technical details.
-    
+
     Learn more about Diátaxis at <https://diataxis.fr/>.
 
 **To learn more about the primary functions:**
@@ -21,4 +21,4 @@ The [Search](./explanation/search.md) section introduces how to search for NASA 
 
 The [Access](./explanation/access.md) section introduces how to use `earthaccess` to download data to a local machine or stream data by opening it directly in memory.
 
-For those unfamiliar, the [NASA glossary](./reference/glossary/nasa-glossary.md) and [cloud computing terminology](./reference/glossary/cloud-glossary.md) pages provide helpful vocabulary for working with NASA Earth science data. 
+For those unfamiliar, the [NASA glossary](./reference/glossary/nasa-glossary.md) and [cloud computing terminology](./reference/glossary/cloud-glossary.md) pages provide helpful vocabulary for working with NASA Earth science data.
