### 2026-07-01 a468ee2
Bump the gha-dependencies group with 4 updates
Bumps the gha-dependencies group with 4 updates: [actions/checkout](https://github.com/actions/checkout), [codecov/codecov-action](https://github.com/codecov/codecov-action), [github-community-projects/issue-metrics](https://github.com/github-community-projects/issue-metrics) and [astral-sh/setup-uv](https://github.com/astral-sh/setup-uv).


Updates `actions/checkout` from 6.0.2 to 7.0.0
- [Release notes](https://github.com/actions/checkout/releases)
- [Changelog](https://github.com/actions/checkout/blob/main/CHANGELOG.md)
- [Commits](https://gith
[…]
