+++ .github/workflows/zizmor.yml
@@ -20,7 +20,7 @@ jobs:
       security-events: "write" # Required for upload-sarif (used by zizmor-action) to upload SARIF files.
     steps:
       - name: "Checkout repository"
-        uses: "actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd" # v6.0.2
+        uses: "actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0" # v7.0.0
         with:
           persist-credentials: false
 
