+++ .github/actions/install-pkg/action.yml
@@ -14,7 +14,7 @@ runs:
   using: composite
   steps:
     - name: Install uv
-      uses: astral-sh/setup-uv@08807647e7069bb48b6ef5acd8ec9567f424441b # v8.1.0
+      uses: astral-sh/setup-uv@fac544c07dec837d0ccb6301d7b5580bf5edae39 # v8.2.0
       with:
         enable-cache: True
         cache-suffix: ${{ inputs.cache-key}}-${{ inputs.python-version }}
