### 2026-07-01 b2e1496
Fix incorrect version comment dependabot screwed up
¯\_(ツ)_/¯

Co-authored-by: Matt Fisher <3608264+mfisher87@users.noreply.github.com>
