### 2026-07-01 a83b5d5
fix RTD bullet formatting
