### 2026-07-01 3af6ba0
Re-label lifecycle detail sections
