### 2026-07-01 3653f4a
reorg triaging sections and content
