### 2026-07-01 83ac905
Update new issue guidance
