+++ docs/contributor/triaging.md
@@ -15,7 +15,9 @@ Details on each of these workflow steps are provided below.
 
 ### A new issue is created
 
+Issues are created using one of several templates or as a blank issue. See the issue template choices [here](https://github.com/earthaccess-dev/earthaccess/issues/new/choose). When an issue is first created, provide initial acknowledgement and gratitude for the submission as a text or emoji response. 
 
+By default, all new issues are created without a project status. Issues without a status are listed in the [Needs Triage](https://github.com/orgs/earthaccess-dev/projects/1/views/3) project view. 
 
 ### The issue is triaged
 
@@ -26,16 +28,16 @@ Triaging is led by the `earthaccess` community manager, though any community mem
 - Adding an issue prioritization.
 - Responding and follow up as needed (i.e. tagging relevant earthaccess maintainers for further support).
 
-By default, all new issues are created without a project status. Issues without a status are listed in the [Needs Triage](https://github.com/orgs/earthaccess-dev/projects/1/views/3) project view. Statuses include:
+#### Moving an issue to backlog status
+
+When triaging a new issue, review the information and provide a response or follow up with question(s) if needed. Move the issue to the Backlog unless it ought to be closed as "not planned", as outlined below. If you are unsure, add the **needs: triage** label. On the righthand side of the issue page, the "Projects" section contains an `earthaccess` project box. Click "no status" to select the status options. Statuses include:
 
 - Backlog
 - In Progress
 - In Review
 - Done
 
-#### Moving an issue to backlog status
-
-When triaging a new issue, review the information and provide a response or follow up with question(s) if needed. Provide gratitude for the submission as a text or emoji response. Move the issue to the Backlog unless it ought to be closed as "not planned", as outlined below. If you are unsure, add the **needs: triage** label. On the righthand side of the issue page, the "Projects" section contains an `earthaccess` project box. Click "no status" to select the status options. Select "Backlog". This will move the project out of the [Needs Triage](https://github.com/orgs/earthaccess-dev/projects/1/views/3) project view into its relevant backlog view depending on issue type. See below for more details on these other project views. 
+Select "Backlog". This will move the project out of the [Needs Triage](https://github.com/orgs/earthaccess-dev/projects/1/views/3) project view into its relevant backlog view depending on issue type. See below for more details on these other project views. 
 
 #### When to "Close as not planned"?
 
