### 2026-07-01 213e1d5
Remove label descriptions
