### 2026-07-01 56ca933
Update grooming guidance
