### 2026-07-02 047504a
Increase workflow headers so TOC shows up
