### 2026-07-03 e08f588
Add DataGranule.crs and document footprint-vs-data CRS distinction
CMR granule footprints (and the GeoJSON from `__geo_interface__`) are always
WGS84 lon/lat, but the underlying data are frequently gridded in a projected
CRS (Sinusoidal, Polar Stereographic, UTM). Nothing surfaced this, so users
could silently assume the footprint CRS matched the data CRS.

Add a `DataGranule.crs` property that returns "EPSG:4326" when the granule has
a horizontal footprint (else None), update the `__geo_interface__` docstring
and the spatial search guide to state that the footprint CRS is not guaranteed
to ma
[…]
