### 2026-07-06 c4c1b9f
chore: update pre-commit hooks
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.15.4 → v0.15.20](https://github.com/astral-sh/ruff-pre-commit/compare/v0.15.4...v0.15.20)
- [github.com/astral-sh/uv-pre-commit: 0.10.7 → 0.11.26](https://github.com/astral-sh/uv-pre-commit/compare/0.10.7...0.11.26)
- [github.com/codespell-project/codespell: v2.4.1 → v2.4.2](https://github.com/codespell-project/codespell/compare/v2.4.1...v2.4.2)
- [github.com/shellcheck-py/shellcheck-py: 1de4a94008448d1ffa57080fbf183feb615ee165 → v0.11.0.1](https://github.com/shellcheck-py/shellcheck-py/compare/1de4a94008448d1f
[…]
