### 2026-07-07 e0ea7b7
Update nav to remove obsolete/ empty how-tos
