+++ docs/.nav.yml
@@ -28,13 +28,10 @@ nav:
       - "Quick start": "user/quick-start.md"
       - How-to:
           - "Authenticate with Earthdata Login": "user/howto/authenticate.md"
-          - "Search NASA datasets using filters": "user/howto/search-collections.md"
           - "Search for data using filters": "user/howto/search-granules.md"
-          - "Determine if a dataset is available in the cloud": "user/howto/storage-location.md"
           - "Access data": "user/howto/access-data.md"
           - "Using authenticated sessions to access data": "user/howto/edl.ipynb"
           - "Remote access and caching with earthaccess": "user/howto/fsspec.ipynb"
-          - "Download data from on-prem location": "user/howto/onprem.md"
           - "Direct S3 access - Open/stream files in the cloud": "user/howto/cloud.md"
           - "Search for services": "user/howto/search-services.md"
       - Tutorials:
