### 2026-07-07 7301ba5
Delete docs/user/howto/search-collections.md
