+++ docs/user/howto/search-collections.md
@@ -1,3 +1,0 @@
-# how to search for data collections using spatial, temporal,  keyword filters
-
-Coming soon
