+++ docs/contributor/triaging.md
@@ -4,7 +4,8 @@ This document outlines our approach to triaging issues in GitHub, including guid
 
 **We hope that this guide will empower anyone to contribute to issue triaging, and address a common question for contributors: "I'm interested in working on the highest priority issues that will solve important problems facing the `earthaccess` community. Where do I begin?"**
 
-See {ref}`issue-grooming` below for details on how to prioritize `earthaccess` issues and browse these via the [`earthaccess` GitHub Project](https://github.com/orgs/earthaccess-dev/projects/1).
+See [Jump to it](#issue-grooming) below for details on how to prioritize `earthaccess` issues and browse these via the [`earthaccess` GitHub Project](https://github.com/orgs/earthaccess-dev/projects/1).
+
 
 ## Issue Lifecycle
 
@@ -74,7 +75,7 @@ To link to a label in GitHub Markdown, copy-and-paste the URL to the label by ri
 https://github.com/earthaccess-dev/earthaccess/labels/good%20first%20issue
 ```
 
-(issue-grooming)=
+<a name="issue-grooming"></a>
 ## The issue is groomed
 
 TODO: Under the groomed section, have more of an intro or sub section on prioritizing issues. Then we can introduce the project views and how they are used, thinking about the contributor audience in addition to triager.
