### 2026-07-07 dee3985
Add project view details
