### 2026-07-08 ff271d9
Update broken link in pull_request_template.md
Updated from https://earthaccess.readthedocs.io/en/latest/contributor/howto/pr-guide/ to https://earthaccess.readthedocs.io/en/latest/contributor/howto/pr-guide/
