+++ docs/user/howto/storage-location.md
@@ -1,3 +1,0 @@
-# Determine if a data set is on-prem or in the cloud
-
-Coming soon
