### 2026-07-09 4421509
Add `asf_search` entry
