### 2026-07-21 8e915ef
add stubs for howto reorg
