+++ docs/user/howto/get_s3_credentials.md
@@ -1,0 +1,1 @@
+# How do I get AWS S3 credentials?
