+++ docs/user/howto/authenticate_netrc.md
@@ -1,0 +1,1 @@
+# How do I authenticate with Earthdata Login using a `.netrc`
