+++ docs/user/howto/authenticate_envar.md
@@ -1,0 +1,1 @@
+# How do I authenticate Earthdata Login using Environment Variables
