### 2026-07-24 1513030
add content to get_s3_credentials
