### 2026-08-01 2b5366a
Bump the gha-dependencies group with 8 updates
Bumps the gha-dependencies group with 8 updates:

| Package | From | To |
| --- | --- | --- |
| [actions/checkout](https://github.com/actions/checkout) | `7.0.0` | `7.0.1` |
| [jupyterlab/maintainer-tools/.github/actions/base-setup](https://github.com/jupyterlab/maintainer-tools) | `0.34.1` | `0.34.2` |
| [jupyterlab/maintainer-tools/.github/actions/check-links](https://github.com/jupyterlab/maintainer-tools) | `0.34.1` | `0.34.2` |
| [tiangolo/issue-manager](https://github.com/tiangolo/issue-manager) | `0.6.0` | `0.8.1` |
| [github-community-proj
[…]
