+++ .github/workflows/zizmor.yml
@@ -20,9 +20,9 @@ jobs:
       security-events: "write" # Required for upload-sarif (used by zizmor-action) to upload SARIF files.
     steps:
       - name: "Checkout repository"
-        uses: "actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0" # v7.0.0
+        uses: "actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1" # v7.0.1
         with:
           persist-credentials: false
 
       - name: "Run zizmor 🌈"
-        uses: "zizmorcore/zizmor-action@5f14fd08f7cf1cb1609c1e344975f152c7ee938d" # v0.5.6
+        uses: "zizmorcore/zizmor-action@6599ee8b7a49aef6a770f63d261d214911a7ce02" # v0.6.0
