+++ .github/actions/install-pkg/action.yml
@@ -14,14 +14,14 @@ runs:
   using: composite
   steps:
     - name: Install uv
-      uses: astral-sh/setup-uv@fac544c07dec837d0ccb6301d7b5580bf5edae39 # v8.2.0
+      uses: astral-sh/setup-uv@11f9893b081a58869d3b5fccaea48c9e9e46f990 # v8.3.2
       with:
         enable-cache: True
         cache-suffix: ${{ inputs.cache-key}}-${{ inputs.python-version }}
 
     # Typically faster than using `uv python install` because of GitHub's
     # cache. See: https://docs.astral.sh/uv/guides/integration/github/#setting-up-python
-    - uses: actions/setup-python@a309ff8b426b58ec0e2a45f0f869d46889d02405 # v6
+    - uses: actions/setup-python@ece7cb06caefa5fff74198d8649806c4678c61a1 # v6
       with:
         python-version: ${{ inputs.python-version }}
 
