### 2026-08-02 earthaccess-dev/earthaccess#1412 @Sherwin-14
Improve README
Raised during PyOpenSci [review](https://github.com/pyOpenSci/software-submission/issues/302#issuecomment-5064346023). These are small README gaps noted by the reviewer:

- [ ] Document the EDL account creation step in the README.
- [ ] Add a copy-paste citation section to the README.
- [ ] Add a CI status badge.
