### 2026-08-02 earthaccess-dev/earthaccess#1412 @Sherwin-14
Created in error closing this out. Apologies for the noise!
