### 2026-08-03 earthaccess-dev/earthaccess#1412 @mfisher87
These do seem like useful problems to solve. Curious why you closed the issue? 🤔
