### 2026-08-03 3e1ad41
Run zizmor on changes to actions in the `actions/` directory
