### 2026-08-07 earthaccess-dev/earthaccess#1412 @Sherwin-14
@mfisher87 I was thinking of creating a master issue (related to PyOpenSci Review) and then create sub issues within that one master issue.
