### 2026-08-07 earthaccess-dev/earthaccess#1412 @mfisher87
Sick, great idea! Thanks for taking that on ❤️
