### 2026-08-08 earthaccess-dev/earthaccess#1423 @Sherwin-14
Resolve PyOpenSci Review Action Items
This issue tracks action items coming out of the  [review](https://github.com/pyOpenSci/software-submission/issues/302#issuecomment-5064903597) of `earthaccess`, which has now been completed.

A few review comments are clarifying questions rather than action items, I haven't added them here since they can be tackled directly in the conversation thread. Should I still list them here, or is it fine to leave them in the thread, since some of them could prompt further discussion?

cc: @mfisher87 @danielfromearth @jhkennedy @betolink
