### 2026-08-08 earthaccess-dev/earthaccess#1423 @mfisher87
🙇 Thank you Sherwin!
