### 2026-08-10 earthaccess-dev/earthaccess#1423 @asteiker
This is so helpful @Sherwin-14 , thank you!! I also just linked #1301 (with an open PR! 🙌 ) based on [this comment](https://github.com/pyOpenSci/software-submission/issues/302#issuecomment-5064346023).
