### 2026-08-17 86a9b90
Add icesat2 ref, and remove harmony-py - per code review
Co-authored-by: Daniel Kaufman <114174502+danielfromearth@users.noreply.github.com>
Co-authored-by: Jessica Scheick <JessicaS11@users.noreply.github.com>
