+++ paper/paper.bib
@@ -193,3 +193,16 @@
   publisher = {American Geophysical Union (AGU)},
   doi = {10.1029/2025JC023511}
 }
+
+@article{icesat2,
+title = {The {Ice, Cloud, and land Elevation Satellite-2 (ICESat-2)}: Science requirements, concept, and implementation},
+journal = {Remote Sensing of Environment},
+volume = {190},
+pages = {260-273},
+year = {2017},
+issn = {0034-4257},
+doi = {10.1016/j.rse.2016.12.029},
+url = {https://www.sciencedirect.com/science/article/pii/S0034425716305089},
+author = {Thorsten Markus and Tom Neumann and Anthony Martino and Waleed Abdalati and Kelly Brunt and Beata Csatho and Sinead Farrell and Helen Fricker and Alex Gardner and David Harding and Michael Jasinski and Ron Kwok and Lori Magruder and Dan Lubin and Scott Luthcke and James Morison and Ross Nelson and Amy Neuenschwander and Stephen Palm and Sorin Popescu and CK Shum and Bob E. Schutz and Benjamin Smith and Yuekui Yang and Jay Zwally},
+keywords = {ICESat-2, Land ice, Sea ice, Vegetation, Climate change, Satellite mission},
+}
