### 2026-08-18 f3a71e2
Migrate all attribute-like methods of DataGranule and DataCollection to @property
And rename `DataCollection.get_links` -> `DataCollection.data_links`
