+++ earthaccess/results.py
@@ -1,6 +1,5 @@
 import json
 import uuid
-import warnings
 from functools import cache
 from typing import Any, ClassVar
 
@@ -89,6 +88,7 @@ class DataCollection(CustomDict):
         "DirectDistributionInformation",
     ]
 
+    @property
     def summary(self) -> dict[str, Any]:
         """Summary containing short_name, concept-id, file-type, and cloud-info (if cloud-hosted).
 
@@ -96,29 +96,16 @@ class DataCollection(CustomDict):
             A summary of the collection metadata.
         """
         # we can print only the concept-id
-
-        warnings.warn(
-            "As of version 1.0, `DataCollection.summary` will be accessed as an "
-            "attribute; e.g. use `DataCollection.summary` **not** "
-            "`DataCollection.summary()`",
-            category=FutureWarning,
-            stacklevel=2,
-        )
-
-        # Only issue a single FutureWarning
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore", category=FutureWarning)
-
-            summary_dict: dict[str, Any]
-            summary_dict = {
-                "short-name": self.get_umm("ShortName"),
-                "concept-id": self.concept_id(),
-                "version": self.version(),
-                "file-type": self.data_type(),
-                "get-data": self.get_data(),
-            }
-            if "Region" in self.s3_bucket():
-                summary_dict["cloud-info"] = self.s3_bucket()
+        summary_dict: dict[str, Any]
+        summary_dict = {
+            "short-name": self.get_umm("ShortName"),
+            "concept-id": self.concept_id(),
+            "version": self.version(),
+            "file-type": self.data_type(),
+            "get-data": self.get_data(),
+        }
+        if "Region" in self.s3_bucket():
+            summary_dict["cloud-info"] = self.s3_bucket()
 
         return summary_dict
 
@@ -133,20 +120,13 @@ class DataCollection(CustomDict):
         """
         return self["umm"].get(umm_field, "")
 
+    @property
     def doi(self) -> str | None:
         """Retrieve the Digital Object Identifier (DOI) for this collection.
 
         Returns:
             This collection's DOI information, or `None`, if it has none.
         """
-        warnings.warn(
-            "As of version 1.0, `DataCollection.doi` will be accessed as an "
-            "attribute; e.g. use `DataCollection.doi` **not** "
-            "`DataCollection.doi()`",
-            category=FutureWarning,
-            stacklevel=2,
-        )
-
         doi = self["umm"].get("DOI", {})
         if isinstance(doi, dict):
             return doi.get("DOI", None)
@@ -173,133 +153,77 @@ class DataCollection(CustomDict):
             else _citation(doi=doi, format_=format, language=language)
         )
 
+    @property
     def concept_id(self) -> str:
         """Return the collection concept ID.
 
         Returns:
             A collection's `concept_id`.This id is the most relevant search field on granule queries.
         """
-        warnings.warn(
-            "As of version 1.0, `DataCollection.concept_id` will be accessed as an "
-            "attribute; e.g. use `DataCollection.concept_id` **not** "
-            "`DataCollection.concept_id()`",
-            category=FutureWarning,
-            stacklevel=2,
-        )
-
         return self["meta"]["concept-id"]
 
+    @property
     def data_type(self) -> str:
         """Return the collection file distribution type.
 
         Returns:
             The collection data type, i.e. HDF5, CSV etc., if available.
         """
-        warnings.warn(
-            "As of version 1.0, `DataCollection.data_type` will be accessed as an "
-            "attribute; e.g. use `DataCollection.data_type` **not** "
-            "`DataCollection.data_type()`",
-            category=FutureWarning,
-            stacklevel=2,
-        )
-
         return str(
             self["umm"]
[…]
