### 2026-08-18 ba72d64
Phrasing
