### 2026-08-18 9d92536
Fixup calls to new property methods
