+++ earthaccess/store.py
@@ -524,7 +524,7 @@ class Store:
         pqdm_kwargs: Mapping[str, Any] | None = None,
         open_kwargs: dict[str, Any] | None = None,
     ) -> list[Any]:
-        total_size = round(sum([granule.size() for granule in granules]) / 1024, 2)
+        total_size = round(sum([granule.size for granule in granules]) / 1024, 2)
         logger.info(
             "Opening %s granules, approx size: %s GB",
             len(granules),
@@ -840,7 +840,7 @@ class Store:
                 for granule in granules
             ),
         )
-        total_size = round(sum(granule.size() for granule in granules) / 1024, 2)
+        total_size = round(sum(granule.size for granule in granules) / 1024, 2)
         logger.info(
             "Getting %s granules, approx download size: %s GB",
             len(granules),
