### 2026-08-18 6f8a432
Fixup more calls to properties
