+++ tests/unit/test_services.py
@@ -51,7 +51,7 @@ class TestServices(VCRTestCase):
         earthaccess._auth.authenticated = False
 
         assert len(datasets) > 0
-        results = datasets[0].services()
+        results = datasets[0].services
 
         self.assertTrue(
             results["S2004184019-POCLOUD"][0]["meta"]["provider-id"] == "POCLOUD",
