### 2026-08-18 8877b32
More property calls
