+++ tests/integration/test_services.py
@@ -33,7 +33,7 @@ class TestServices(VCRTestCase):
         )
 
         dataset_services = {
-            dataset["umm"]["ShortName"]: dataset.services() for dataset in datasets
+            dataset["umm"]["ShortName"]: dataset.services for dataset in datasets
         }
 
         self.assertEqual(next(iter(dataset_services.keys())), "MUR-JPL-L4-GLOB-v4.1")
