### 2026-08-18 #302 @kysolvik
Thanks @ocefpaf , @stefano-m-potter, and @danielfromearth! I've updated the status to "reviews-in-awaiting-changes".

Once the changes have been made and `earthaccess` is ready for the reviewers to take another look at, let us know. Usually we aim for ~2 weeks for the changes, but let me know if you need more time @danielfromearth.
