### 2026-08-19 4d8e3d8
Update data access workflow description and steps per code review
Co-authored-by: Daniel Kaufman <114174502+danielfromearth@users.noreply.github.com>
