### 2026-08-20 ddd8754
Improve readability of changelog
Co-authored-by: Chuck Daniels <chuck@developmentseed.org>
