### 2026-08-20 bcbb398
Add detail about new way to access fields
