### 2026-08-20 5c0587e
Revert granule.data_links to non-property, make granule.dataviz_links a property
