+++ earthaccess/formatters.py
@@ -34,7 +34,7 @@ def _repr_granule_html(granule: Any) -> str:
     data_links = "".join(
         [
             f'<a href="{link}" target="_blank" class="btn btn-secondary btn-sm">{link.split("/")[-1]}</a>'
-            for link in granule.data_links
+            for link in granule.data_links()
         ],
     )
     granule_size = round(granule.size, 2)
