+++ CHANGELOG.md
@@ -27,15 +27,18 @@ and this project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
   - `DataCollection.services()` -> `DataCollection.services`
   - `DataCollection.summary()` -> `DataCollection.summary`
   - `DataCollection.version()` -> `DataCollection.version`
+  - `DataGranule.dataviz_links()` -> `DataGranule.dataviz_links`
   - `DataGranule.size()` -> `DataGranule.size`
 
   **NOTE:** The method `DataCollection.get_links()` was replaced with the field
-  `DataCollection.data_links`. The method name was `get_links` because in the
-  CMR API, the links are described as links of type "GET DATA".  However, while
-  the old method name matches the name of the link type in the CMR API, it can
-  be read as a verb ("get" the links, rather than links of _type_ "get"), when
-  it's an attribute (noun).  Thus, the name `data_links` was chosen to avoid
-  confusion (noun instead of verb).
+  `DataCollection.data_links`. The method name was previously `get_links` because in
+  the CMR API, the links are described as links of type "GET DATA". However, while the
+  old method name matches the name of the link type in the CMR API, it can be read as a
+  verb ("get" the links, rather than links of _type_ "get"), when it's an attribute
+  (noun). The name `data_links` was chosen to avoid confusion (noun instead of verb),
+  and for consistency with `DataGranule.data_links()` (SUBNOTE:
+  `DataGranule.data_links()` has arguments, so it was **not** changed to a property at
+  this time).
 
 ### Fixed
 
