### 2026-08-20 c0f53cc
Fixup call to property
