+++ docs/user/explanation/backwards-compatibility.md
@@ -43,6 +43,39 @@ Our project follows the [SPEC0](https://scientific-python.org/specs/spec-0000/) 
 
 ## Migration guides
 
+### 0.19.0
+
+* Many `DataCollection` and `DataGranule` methods are now read-only fields.
+  For example, you'll get the size of a `DataGranule` named `granule` via
+  `granule.size` instead of `granule.size()`. If you use the old syntax, you'll
+  receive errors like `TypeError: 'NoneType' object is not callable`,
+  `TypeError: 'str' object is not callable`, etc.
+
+  The following methods were changed to read-only fields of the same name,
+  except where noted:
+
+  * `DataCollection.concept_id()` -> `DataCollection.concept_id`
+  * `DataCollection.data_type()` -> `DataCollection.data_type`
+  * `DataCollection.doi()` -> `DataCollection.doi`
+  * `DataCollection.get_links()` -> `DataCollection.data_links` (note the name change, details below)
+  * `DataCollection.landing_page()` -> `DataCollection.landing_page`
+  * `DataCollection.s3_bucket()` -> `DataCollection.s3_bucket`
+  * `DataCollection.services()` -> `DataCollection.services`
+  * `DataCollection.summary()` -> `DataCollection.summary`
+  * `DataCollection.version()` -> `DataCollection.version`
+  * `DataGranule.dataviz_links()` -> `DataGranule.dataviz_links`
+  * `DataGranule.size()` -> `DataGranule.size`
+
+  **NOTE:** The method `DataCollection.get_links()` was replaced with the field
+  `DataCollection.data_links`. The method name was previously `get_links` because in
+  the CMR API, the links are described as links of type "GET DATA". However, while the
+  old method name matches the name of the link type in the CMR API, it can be read as a
+  verb ("get" the links, rather than links of _type_ "get"), when it's an attribute
+  (noun). The name `data_links` was chosen to avoid confusion (noun instead of verb),
+  and for consistency with `DataGranule.data_links()` (SUBNOTE:
+  `DataGranule.data_links()` has arguments, so it was **not** changed to a property at
+  this time).
+
 ### 0.14.0
 
 * `earthaccess.login()` will now raise an exception if Earthdata Login rejects credentials.
