### 2026-08-20 c93deea
Remove unused function
