+++ earthaccess/formatters.py
@@ -14,10 +14,6 @@ def _load_static_files() -> list[str]:
     ]
 
 
-def _repr_collection_html() -> str:
-    return "<div></div>"
-
-
 def _repr_granule_html(granule: Any) -> str:
     css_styles = _load_static_files()
     css_inline = f"""<div id="{uuid4()}" style="height: 0px; display: none">
