### 2026-08-20 9a401a3
Acquiesce to Mkdocs awful non-standard Markdown behavior
