### 2026-08-20 b8bba67
Add type annotation to _repr_granule_html
