### 2026-08-20 fb3dd10
Fixup incorrect Any annotation
