### 2026-08-20 588db95
Fixup incorrect Any annotation
