### 2026-08-20 004a8d6
Tell Store class that self.auth can be None
