### 2026-08-20 f7eef07
Add type guards
