### 2026-08-24 earthaccess-dev/earthaccess#1423 @asteiker
@danielfromearth @Sherwin-14 It looks like the reviewers are asking for a ~2-week turnaround to address changes. It's not clear to me whether they're expecting us to update everything that's listed in this issue, and whether we also need a new release to reflect those changes. Can you help clarify?
