### 2026-08-25 earthaccess-dev/earthaccess#1423 @Sherwin-14
I do think we will need a new release once we resolve this #1307.
