### 2026-09-01 5998348
Bump the gha-dependencies group with 7 updates
Bumps the gha-dependencies group with 7 updates:

| Package | From | To |
| --- | --- | --- |
| [jupyterlab/maintainer-tools/.github/actions/base-setup](https://github.com/jupyterlab/maintainer-tools) | `0.34.2` | `0.35.0` |
| [jupyterlab/maintainer-tools/.github/actions/check-links](https://github.com/jupyterlab/maintainer-tools) | `0.34.2` | `0.35.0` |
| [hynek/build-and-inspect-python-package](https://github.com/hynek/build-and-inspect-python-package) | `2.18.0` | `3.0.1` |
| [pypa/gh-action-pypi-publish](https://github.com/pypa/gh-action-pypi-
[…]
