+++ .github/workflows/zizmor.yml
@@ -27,4 +27,4 @@ jobs:
           persist-credentials: false
 
       - name: "Run zizmor 🌈"
-        uses: "zizmorcore/zizmor-action@6599ee8b7a49aef6a770f63d261d214911a7ce02" # v0.6.0
+        uses: "zizmorcore/zizmor-action@3dc1ecc9bcb9e94e9b2c709687979e1298497054" # v0.6.2
