### 2026-09-02 #302 @danielfromearth
@kysolvik, thanks for the note and the heads up on time frame. We've been making progress on addressing the reviewers' comments, but I see that we just hit the two week mark. If we could please have until next week, I think we could have a new release with everything addressed.
