### 2026-09-02 earthaccess-dev/earthaccess#1423 @danielfromearth
Note: I think once this issue's sub-issues are in a release, then this issue can be closed.
