### 2026-09-03 f40532a
Bump the gha-dependencies group across 1 directory with 8 updates
Bumps the gha-dependencies group with 8 updates in the / directory:

| Package | From | To |
| --- | --- | --- |
| [jupyterlab/maintainer-tools/.github/actions/base-setup](https://github.com/jupyterlab/maintainer-tools) | `0.34.2` | `0.35.0` |
| [jupyterlab/maintainer-tools/.github/actions/check-links](https://github.com/jupyterlab/maintainer-tools) | `0.34.2` | `0.35.0` |
| [github-community-projects/issue-metrics](https://github.com/github-community-projects/issue-metrics) | `5.0.0` | `5.0.1` |
| [hynek/build-and-inspect-pytho
[…]
