### 2026-09-03 11ee8bb
Connect tool descriptions to "statement of need" workflow steps
Co-authored-by: Joseph H Kennedy <me@jhkennedy.org>

Co-authored-by: Daniel Kaufman <114174502+danielfromearth@users.noreply.github.com>
