### 2026-09-03 c1688fb
Connect tool descriptions to "statement of need" workflow steps
Co-authored-by: Joseph H Kennedy <me@jhkennedy.org>
