### 2026-09-04 bca4605
update AI policy mention
