### 2026-09-04 6cc7bbb
add llm reference to disclosure
