### 2026-09-04 03b5afb
add remark about interoperability with python ecosystem
Co-authored-by: betolink <717735+betolink@users.noreply.github.com>
