### 2026-09-04 bd7a6b6
add remark about improvements for ATL06 and other granule-heavy products
Co-authored-by: betolink <717735+betolink@users.noreply.github.com>
