### 2026-09-04 9a7d2c1
docs(paper): refresh bib entries and submission-date metadata
- Add @software{asf_search} and @misc{tempo_o3tot_l3} references
- Add @misc{atl06} data DOI; cite it (not the mission paper) where the
  text refers to the ATL06 product, so @icesat2 refers only to the mission
- Remove @software{harmony_py} because there's no citation in paper.md
- Update the NSIDC publisher field, to avoid splits on "and" into "National Snow; Ice Data Center"
- Update frontmatter date and dependent-repo count (230 -> 260)
