+++ paper/paper.bib
@@ -5,6 +5,13 @@
   url = {https://github.com/nasa/python_cmr}
 }
 
+@software{asf_search,
+  title = {asf\_search: {Python} wrapper for the {Alaska Satellite Facility (ASF)} {SearchAPI}},
+  author = {{Alaska Satellite Facility Discovery Team}},
+  year = {2026},
+  url = {https://github.com/asfadmin/Discovery-asf_search}
+}
+
 @software{fsspec,
   title = {filesystem\_spec: Filesystem interfaces for {Python}},
   author = {Durant, Martin and others},
@@ -117,13 +124,6 @@
   doi = {10.21105/joss.04912}
 }
 
-@software{harmony_py,
-  title = {harmony-py: {NASA Harmony Python} Client},
-  author = {{NASA}},
-  year = {2026},
-  url = {https://github.com/nasa/harmony-py}
-}
-
 @article{xarray,
   title = {xarray: {N-D} labeled arrays and datasets in {Python}},
   author = {Hoyer, Stephan and Hamman, Joe},
@@ -205,4 +205,24 @@ doi = {10.1016/j.rse.2016.12.029},
 url = {https://www.sciencedirect.com/science/article/pii/S0034425716305089},
 author = {Thorsten Markus and Tom Neumann and Anthony Martino and Waleed Abdalati and Kelly Brunt and Beata Csatho and Sinead Farrell and Helen Fricker and Alex Gardner and David Harding and Michael Jasinski and Ron Kwok and Lori Magruder and Dan Lubin and Scott Luthcke and James Morison and Ross Nelson and Amy Neuenschwander and Stephen Palm and Sorin Popescu and CK Shum and Bob E. Schutz and Benjamin Smith and Yuekui Yang and Jay Zwally},
 keywords = {ICESat-2, Land ice, Sea ice, Vegetation, Climate change, Satellite mission},
+}
+
+@misc{atl06,
+  title = {{ATLAS/ICESat-2 L3A} Land Ice Height, {Version 7}},
+  author = {Smith, Ben and Adusumilli, Susheel and Csath{\'o}, Be{\'a}ta and Felikson, Denis and Fricker, Helen and Gardner, Alex and Holschuh, Nick and Lee, Jeff and Nilsson, Johan and Paolo, Fernando and Siegfried, Matthew and Sutterley, Tyler and {the ICESat-2 Science Team}},
+  year = {2025},
+  publisher = {{NASA National Snow and Ice Data Center Distributed Active Archive Center}},
+  doi = {10.5067/ATLAS/ATL06.007},
+  url = {https://doi.org/10.5067/ATLAS/ATL06.007},
+  note = {Accessed: 2026-09-04}
+}
+
+@misc{tempo_o3tot_l3,
+  title = {{TEMPO} Gridded Ozone Total Column {V04} ({PROVISIONAL})},
+  author = {Liu, Xiong},
+  year = {2026},
+  publisher = {{NASA} Langley Atmospheric Science Data Center Distributed Active Archive Center},
+  doi = {10.5067/IS-40e/TEMPO/O3TOT_L3.004},
+  url = {https://doi.org/10.5067/IS-40e/TEMPO/O3TOT_L3.004},
+  note = {Accessed: 2026-09-02}
 }
