### 2026-09-04 earthaccess-dev/earthaccess#1423 @danielfromearth
All sub-issues closed 🥳
