### 2026-09-04 #302 @danielfromearth
HI @kysolvik, thank you again for your attention to this submission; I believe all of the remaining issues referenced above have now been addressed in the [0.19.0 release](https://github.com/earthaccess-dev/earthaccess/releases/tag/v0.19.0) made yesterday, including the requested README updates and support for Python 3.14.  When you have a chance, please let me know what you see as the next steps!

(Moreover, we have a [full JOSS manuscript draft](https://github.com/earthaccess-dev/earthaccess/pull/1249) nearing a merge into `main`.)
