### 2026-09-04 #302 @kysolvik
Hi @danielfromearth, thanks for the update! And sorry I didn't reply to your message from a couple days ago.

Next step is for the reviewers to look over the changes and determine if their concerns/comments have been addressed. @ocefpaf and @stefano-m-potter, there's no template for this stage (just a [short description in the guide](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#review-follow-up)). If there's any outstanding items, let us know. Once things look good, we can go ahead with acceptance.

Usually we leave about a week for reviewers to take a look at the changes. Do you think the two of you could take a look at some point next week?

Thanks all!
