### 2026-09-04 f24279e
add doc on visualizing granules
