### 2026-09-05 #302 @ocefpaf
Thanks @kysolvik. Everything is addressed IMO and I would like to thanks @danielfromearth and all the maintainers for taking the all the clarifications on the Software and Documentation they have for Eartchacess.

Shameless plug, we publish recently a demonstration of earthacess use for NOAA scientist at: https://ioos.github.io/ioos_code_lab/content/code_gallery/data_access_notebooks/2026-08-25-PACE-OBIS.html

The earthacess package opens up a myriad data investigation and we hope to see manu in the future.
