### 2026-09-05 06dce59
docs(paper): describe access without region auto-detection
Per accepted governance decision 231, us-west-2 region checking will be avoided in favor of an EAFP approach where users can specify direct S3 or HTTPS.

- Updated to no longer claim runtime region detection or automatic access-path selection: Summary, state of the field, authentication & access layers of the design, and the location-transparency design note.
- Convert the software design layers list from numbered to bulleted.

Co-authored-by: Joseph H Kennedy <me@jhkennedy.org>
