+++ paper/paper.md
@@ -116,9 +116,9 @@ experience.
 `earthaccess` provides a unified, high-level Python interface that reduces this
 workflow to just a few lines of code. The library handles authentication with NASA's
 Earthdata Login (EDL) service [@nasa_edl], exposes NASA's Common Metadata Repository
-(CMR) [@nasa_cmr] for data discovery, and transparently manages data retrieval via
-either HTTPS download or direct S3 access when running in the Amazon Web Services (AWS)
-`us-west-2` region -- where data within NASA's Earthdata Cloud reside. `earthaccess` also supports
+(CMR) [@nasa_cmr] for data discovery, and manages data retrieval through a single
+interface covering both HTTPS download and direct Amazon Web Services (AWS) S3 access
+to data in NASA's Earthdata Cloud. `earthaccess` also supports
 streaming data directly into analysis-ready formats using `fsspec` [@fsspec] and
 constructing virtual Zarr stores from archival formats (e.g., HDF5 and NetCDF4) using
 DMR++ metadata [@dmrpp], powered by VirtualiZarr [@virtualizarr] and kerchunk [@kerchunk], enabling drastic improvements in access performance.
@@ -151,8 +151,8 @@ of NASA Earthdata's infrastructure. No existing tool provides this end-to-end,
 provider-agnostic workflow. Rather than reinventing query or filesystem libraries,
 `earthaccess` composes and extends existing open-source tools and contributes the
 NASA-specific domain knowledge (DAAC configurations, credential endpoints,
-authentication flows, cloud-detection logic) that binds them into a usable
-data access layer.
+authentication flows, and the mapping between HTTPS and direct-S3 access links) that
+binds them into a usable data access layer.
 
 Several tools exist for accessing NASA Earth science data, each addressing only part of
 the workflow, or addressing the full workflow for only a subset of NASA's data holdings:
@@ -183,30 +183,30 @@ the workflow, or addressing the full workflow for only a subset of NASA's data h
 `earthaccess` is organized into four core layers, each encapsulating a distinct
 component of the data access workflow:
 
-1. **Authentication**: Manages the full lifecycle of NASA Earthdata Login credentials,
-   supporting environment variables, `.netrc` files, and interactive prompts. Once
-   authenticated, the library creates HTTP sessions that correctly handle NASA's
-   cross-domain redirects and retrieves and renews temporary AWS S3 credentials for in-region
-   cloud access.
+- **Authentication**: Manages the full lifecycle of NASA Earthdata Login credentials,
+  supporting environment variables, `.netrc` files, and interactive prompts. Once
+  authenticated, the library creates HTTP sessions that correctly handle NASA's
+  cross-domain redirects and retrieves and renews temporary AWS S3 credentials for
+  direct S3 access.
 
-2. **Search**: Extends `python-cmr` with DAAC-aware provider resolution, cloud-hosting
-   detection, and deep-paging support. Query results are wrapped in rich objects that
-   preserve the full metadata response while exposing convenience methods for data
-   links, spatial footprints, and formatted citations.
+- **Search**: Extends `python-cmr` with DAAC-aware provider resolution, cloud-hosting
+  detection, and deep-paging support. Query results are wrapped in rich objects that
+  preserve the full metadata response while exposing convenience methods for data
+  links, spatial footprints, and formatted citations.
 
-3. **Access**: Attempts to detect at runtime whether the process is running within AWS `us-west-2`
-   and automatically selects the optimal access path -- direct S3 reads for in-region
-   access or HTTPS access otherwise. Users can manually specify an access path if needed.  Files can be opened as `fsspec`-compatible
-   file-like objects for streaming into libraries such as xarray [@xarray], or
-   downloaded to disk with parallel, fault-tolerant transfers. Because `earthaccess`
[…]
