### 2026-09-08 4452477
consistent hard line wrapping (whitespace only)
