+++ paper/paper.md
@@ -103,51 +103,73 @@ bibliography: paper.bib
 
 # Summary
 
-`earthaccess` is an open-source Python library that simplifies the discovery, authentication,
-and access of NASA Earth science data. NASA's Earth Observing System Data and Information System
-(EOSDIS) distributes over 100 petabytes of data across 11 Distributed Active Archive Centers
-(DAACs) [@nasa_earthdata], encompassing satellite imagery, climate records, atmospheric
-measurements, and other geospatial datasets critical to Earth science research. Accessing these
-data programmatically has historically required researchers to navigate multiple authentication
-systems, understand provider-specific APIs and protocols, and write substantial boilerplate code
+`earthaccess` is an open-source Python library that simplifies the discovery,
+authentication, and access of NASA Earth science data. NASA's Earth Observing System
+Data and Information System (EOSDIS) distributes over 100 petabytes of data across 11
+Distributed Active Archive Centers (DAACs) [@nasa_earthdata], encompassing satellite
+imagery, climate records, atmospheric measurements, and other geospatial datasets
+critical to Earth science research. Accessing these data programmatically has
+historically required researchers to navigate multiple authentication systems,
+understand provider-specific APIs and protocols, and write substantial boilerplate code
 -- challenges that particularly affect researchers without deep software engineering
 experience.
 
-`earthaccess` provides a unified, high-level Python interface that reduces this
-workflow to just a few lines of code. The library handles authentication with NASA's
-Earthdata Login (EDL) service [@nasa_edl], exposes NASA's Common Metadata Repository
-(CMR) [@nasa_cmr] for data discovery, and manages data retrieval through a single
-interface covering both HTTPS download and direct Amazon Web Services (AWS) S3 access
-to data in NASA's Earthdata Cloud. `earthaccess` also supports
-streaming data directly into analysis-ready formats using `fsspec` [@fsspec] and
-constructing virtual Zarr stores from archival formats (e.g., HDF5 and NetCDF4) using
-DMR++ metadata [@dmrpp], powered by VirtualiZarr [@virtualizarr] and kerchunk [@kerchunk], enabling drastic improvements in access performance.
+`earthaccess` provides a unified, high-level Python interface that reduces this workflow
+to just a few lines of code. The library handles authentication with NASA's Earthdata
+Login (EDL) service [@nasa_edl], exposes NASA's Common Metadata Repository (CMR)
+[@nasa_cmr] for data discovery, and manages data retrieval through a single interface
+covering both HTTPS download and direct Amazon Web Services (AWS) S3 access to data in
+NASA's Earthdata Cloud. `earthaccess` also supports streaming data directly into
+analysis-ready formats using `fsspec` [@fsspec] and constructing virtual Zarr stores
+from archival formats (e.g., HDF5 and NetCDF4) using DMR++ metadata [@dmrpp], powered by
+VirtualiZarr [@virtualizarr] and kerchunk [@kerchunk], enabling drastic improvements in
+access performance.
 
 
 # Statement of need
 
 NASA's Earth science data archive is one of the largest and most diverse collections of
 Earth observation data in the world, used by over ten million researchers, educators,
-and decision-makers globally [@nasa_esds_data_metrics]. However, the complexity of the underlying data infrastructure
[…]
