### 2026-09-09 87d5292
adding geo_interface and to_geopandas to DataCollection
