+++ earthaccess/results.py
@@ -20,6 +20,110 @@ def _citation(*, doi: str, format_: str, language: str) -> str:
     response.raise_for_status()
     return response.text
 
+def _geometry_to_geojson(geometry: dict[str, Any]) -> dict[str, object]:
+    """Convert a UMM-G Geometry to a GeoJSON representation. 
+    It is assumed that the `dict` contains only a single key-value pair,
+    and this property is the result of converting the value of that pair
+    into an equivalent GeoJSON structure, depending on the key, as follows:
+
+    |Key                   |Value converted to GeoJSON|
+    |:---------------------|:-------------------------|
+    |`"Lines"`             |`"MultiLineString"`       |
+    |`"Points"`            |`"MultiPoint"`            |
+    |`"BoundingRectangles"`|`"MultiPolygon"`          |
+    |`"GPolygons"`         |`"MultiPolygon"`          |
+
+    Parameters:
+        geometry: A UMM-G Geometry dictionary.
+    
+    Returns:
+        A GeoJSON representation of the UMM-G Geometry.
+
+    Raises:
+        ValueError: If the UMM-G Geometry is invalid or unsupported.
+    """
+    if "GPolygons" in geometry:
+        return {
+            "type": "MultiPolygon",
+            "coordinates": [
+                [
+                    # Outer ring (counterclockwise)
+                    [
+                        [point["Longitude"], point["Latitude"]]
+                        for point in poly["Boundary"]["Points"]
+                    ],
+                    # Inner rings (clockwise)
+                    *(
+                        [
+                            [point["Longitude"], point["Latitude"]]
+                            # In UMM-G, boundary points are always counterclockwise,
+                            # but GeoJSON wants inner rings to be clockwise, so we
+                            # must reverse the points in ExclusiveZone Boundaries.
+                            for point in reversed(boundary["Points"])
+                        ]
+                        # In UMM-G, ExclusiveZone is optional.
+                        for boundary in poly.get("ExclusiveZone", {}).get(
+                            "Boundaries",
+                            [],
+                        )
+                    ),
+                ]
+                for poly in geometry["GPolygons"]
+            ],
+        }
+
+    if "BoundingRectangles" in geometry:
+        return {
+            "type": "MultiPolygon",
+            "coordinates": [
+                [
+                    [
+                        [
+                            rect["WestBoundingCoordinate"],
+                            rect["SouthBoundingCoordinate"],
+                        ],
+                        [
+                            rect["EastBoundingCoordinate"],
+                            rect["SouthBoundingCoordinate"],
+                        ],
+                        [
+                            rect["EastBoundingCoordinate"],
+                            rect["NorthBoundingCoordinate"],
+                        ],
+                        [
+                            rect["WestBoundingCoordinate"],
+                            rect["NorthBoundingCoordinate"],
+                        ],
+                        [
+                            rect["WestBoundingCoordinate"],
+                            rect["SouthBoundingCoordinate"],
+                        ],
+                    ],
+                ]
+                for rect in geometry["BoundingRectangles"]
+            ],
+        }
+
+    if "Points" in geometry:
+        return {
+            "type": "MultiPoint",
+            "coordinates": [
+                [p["Longitude"], p["Latitude"]] for p in geometry["Points"]
+            ],
+        }
+
+    if "Lines" in geometry:
+        return {
+            "type": "MultiLineString",
+            "coordinates": [
+                [[p["Longitude"], p["Latitude"]] for p in line["Points"]]
+                for line in geometry["Lines"]
+            ],
+        }
+
+    msg = f"Invalid Geometry in granule's horizontal spatial extent: {geometry}"
+    raise ValueError(msg)
+
 
 class CustomDict(dict):
     _basic_umm_fields_: ClassVar[list] = []
[…]
