### 2026-09-09 ea6f54c
docs(paper): apply suggestions on wording and code formatting
- remove "ongoing" since (a) it's mostly done and (b) won't be ongoing for future readers.
- remove residual redundancy from Earthdata Cloud rewording
- backtick VirtualiZarr, kerchunk, and xarray package names throughout

Co-authored-by: Ian Carroll <ian.t.carroll@nasa.gov>
