### 2026-09-09 96f6e46
Update affiliation in paper/paper.md
Co-authored-by: Chuck Daniels <chuck@developmentseed.org>
