### 2026-09-12 3aebcd0
Rename CMR query objects to clarify they are queries
