### 2026-09-12 1fa18e0
Update tests to reference renamed query classes
