### 2026-09-15 69155cd
Grammatical improvements and add missing first name
Co-authored-by: Jessica Scheick <JessicaS11@users.noreply.github.com>
