+++ tests/unit/test_geo_interface.py
@@ -1,5 +1,6 @@
 import pytest
-from earthaccess.results import DataGranule
+from earthaccess.results import DataCollection, DataGranule
+from earthaccess.search import DataCollections, DataGranules
 
 # Mapping from test case identifier to UMMG Geometry with the GeoJSON expected
 # to be the value of the __geo_interface__ property of a DataGranule that
@@ -279,3 +280,194 @@ def test_missing_horizontal_spatial_domain_raises():
 
     with pytest.raises(ValueError):
         _ = granule.__geo_interface__
+
+def _make_granule(test_case: dict[str, object], name: str) -> DataGranule:
+    """returns a DataGranule with geometry and key attributes for testing to_geopandas()"""
+    return DataGranule(
+        {
+            "meta": {"concept-id": f"G-{name}", "provider-id": "PROV"},
+            "umm": {
+                "GranuleUR": name,
+                "SpatialExtent": {
+                    "HorizontalSpatialDomain": {"Geometry": test_case["geometry"]},
+                },
+                "TemporalExtent": {
+                    "RangeDateTime": {
+                        "BeginningDateTime": "2024-01-01T00:00:00.000Z",
+                        "EndingDateTime": "2024-01-02T23:59:59.000Z",
+                    },
+                },
+                "RelatedUrls": [
+                    {
+                        "URL": f"https://data.ornldaac.nasa.gov/protected/{name}.nc",
+                        "Type": "GET DATA",
+                        "Description": f"Download {name}.nc",
+                    },
+                ],
+            },
+        },
+    )
+
+
+def test_to_geopandas_returns_geodataframe():
+    gpd = pytest.importorskip("geopandas")
+
+    granules = [
+        _make_granule(TEST_CASES["multiple-gpolygons"], "poly"),
+        _make_granule(TEST_CASES["points"], "point"),
+        _make_granule(TEST_CASES["lines"], "line"),
+    ]
+
+    gdf = DataGranules.to_geopandas(granules)
+
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert str(gdf.crs).upper() == "EPSG:4326"
+    assert len(gdf) == 3
+    assert list(gdf.geometry.geom_type) == [
+        "MultiPolygon",
+        "MultiPoint",
+        "MultiLineString",
+    ]
+    assert gdf.geometry.notna().all()
+
+
+def test_to_geopandas_includes_attribute_columns():
+    pytest.importorskip("geopandas")
+
+    granules = [_make_granule(TEST_CASES["bounding-rectangles"], "rect")]
+    gdf = DataGranules().to_geopandas(granules)
+
+    row = gdf.iloc[0]
+    assert row["concept_id"] == "G-rect"
+    assert row["granule_name"] == "rect"
+    assert row["beginning_datetime"] == "2024-01-01T00:00:00.000Z"
+    assert row["ending_datetime"] == "2024-01-02T23:59:59.000Z"
+    assert row["data_links"] == ["https://data.ornldaac.nasa.gov/protected/rect.nc"]
+    assert "size_MB" in gdf.columns
+
+
+def test_to_geopandas_includes_full_umm_column():
+    pytest.importorskip("geopandas")
+
+    granule = _make_granule(TEST_CASES["points"], "point")
+    gdf = DataGranules().to_geopandas([granule])
+
+    row = gdf.iloc[0]
+    assert row["umm"] == granule["umm"]
+    assert row["umm"]["SpatialExtent"]["HorizontalSpatialDomain"]["Geometry"] == (
+        TEST_CASES["points"]["geometry"]
+    )
+
+
[…]
