### 2026-09-17 3e54514
code cleanups
