### 2026-09-17 fbfe7df
more code cleanups
