### 2026-09-18 e005b13
cleanups with uv precommit
