### 2026-09-21 f7e346c
split affiliations and update numbering
Co-authored-by: Daniel Kaufman <114174502+danielfromearth@users.noreply.github.com>
Co-authored-by: Jessica Scheick <JessicaS11@users.noreply.github.com>
