### 2026-09-22 2fa8232
Update affiliation index numbers
