### 2026-09-26 e0f8827
Use colons in callout list
